server/requirements.txt
flask
flask-cors
llama-index>=0.9.0
transformers>=4.36.0
torch>=2.0.0
huggingface_hub
langchain
langchain-ollama
langchain-chroma
pandas
bitsandbytes>=0.41.0

python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8 # For embeddings
unstructured==0.14.4 # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb==0.5.0 # Vector storage
openai==1.31.1 # For embeddings
tiktoken==0.7.0  # For embeddings 

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 
# Flask and web server
#flask==2.3.3
#flask-cors==4.0.0
werkzeug==2.3.7

# File processing
PyMuPDF==1.23.7  # for PDF processing
docx2txt==0.8
python-docx==1.0.1



#sentence-transformers==2.2.2


tqdm==4.66.1
